add check smtp and add loging.
