+++ WorkerServiceEmail/WorkerServiceEmail/Services/ClientSmtp/SmtpClientYandexAsync.cs
@@ -1,5 +1,6 @@
 ﻿using MailKit.Net.Smtp;
 using MimeKit;
+using WorkerServiceEmail.EntityMessage;
 
 namespace WorkerServiceEmail.Email.SMTP.Client
 {
@@ -24,6 +25,25 @@ namespace WorkerServiceEmail.Email.SMTP.Client
             catch (Exception ex)
             {
                 return false;
+            }
+        }
+
+        public async Task<OutputStatusSmtp> StatusSmtpConnectAsync()
+        {
+            try
+            {
+                using (var client = new SmtpClient())
+                {
+                    await client.ConnectAsync("smtp.yandex.com", 25, false);
+                    await client.AuthenticateAsync(_login, _password);
+                    await client.DisconnectAsync(true);
+
+                    return new OutputStatusSmtp { SmtpServer = "smtp.yandex.com", Status = true };
+                }
+            }
+            catch (Exception ex)
+            {
+                return new OutputStatusSmtp { SmtpServer = "smtp.yandex.com", Status = false, ErrorMessage = ex.Message };
             }
         }
     }
