--- WorkerServiceEmail/WorkerServiceEmail/Services/Interfaces/IClientSmtpConnect.cs
@@ -1,9 +1,0 @@
-﻿using WorkerServiceEmail.EntityMessage;
-
-namespace WorkerServiceEmail.Services.ClientSmtp
-{
-    public interface IClientSmtpConnect
-    {
-        Task<OutputStatusSmtp> StatusSmtpConnectAsync();
-    }
-}
